Updated request methods, created network notifiers and worked on functionality in the main view.

+++ ReignDesignTest/Library/RequestHelper.swift
@@ -8,27 +8,51 @@
 
 import UIKit
 import Alamofire
+import ReachabilitySwift
 
 class RequestHelper: NSObject {
     
-    class func getNewsByDate(completion: () -> Void ){
-        
-        Alamofire.request(.GET, GlobalConstants.ServerAddress.address)
-            .responseJSON { response in
-                //Delete all previous saved entries
-                Entry .deleteAllEntries()
-                //Now get the response
-                let responseDictionary = response.2.value as! NSDictionary
-                //After that, get all the entries returned by the server
-                let entriesDictionary = responseDictionary.objectForKey("hits") as! NSArray
-                
-                //Iterate over the dictionary to create all the entries
-                for entryDictionary in entriesDictionary{
-                    let entry = Entry()
-                    entry.saveEntryWithDictionary(entryDictionary as! NSDictionary)
-                }
-                
-                completion()
+    let reachability = Reachability.reachabilityForInternetConnection()
+    var isNetworkAvailable = true
+    
+    func getNewsByDate(success: () -> Void, failure: () -> Void){
+        //First check if  network is reachable
+        if isNetworkAvailable {
+            Alamofire.request(.GET, GlobalConstants.ServerAddress.address)
+                .responseJSON { response in
+                    //Delete all previous saved entries
+                    Entry .deleteAllEntries()
+                    //Now get the response
+                    let responseDictionary = response.2.value as! NSDictionary
+                    //After that, get all the entries returned by the server
+                    let entriesDictionary = responseDictionary.objectForKey("hits") as! NSArray
+                    
+                    //Iterate over the dictionary to create all the entries
+                    for entryDictionary in entriesDictionary{
+                        let entry = Entry()
+                        entry.saveEntryWithDictionary(entryDictionary as! NSDictionary)
+                    }
+                    success()
+            }
+        }else{
+            failure()
         }
+    }
+    
+    func startNetworkNotifier(){
+        self.isNetworkAvailable = reachability!.isReachable()
+        
+        reachability!.whenReachable = { reachability in
+            self.isNetworkAvailable = true
+        }
+        reachability!.whenUnreachable = { reachability in
+            self.isNetworkAvailable = false
+        }
+        
+        reachability!.startNotifier()
+    }
+    
+    func stopNetworkNotifier(){
+        reachability!.stopNotifier()
     }
 }
